Starting x86 support for Zynthian
Create Zynthian install for use on x86 hardware
Debian 13 packages

Switch to base Debian install, and update arch name
x86_64 is now amd64 in most context.
Refresh git repo if we did not clone it
Was missing updates when switching between ZIP file intial clone
and existing clone from failed install

Package Changes for x86

cpufrequtils is now gone.
Fixup ALSA install
KXStudio provides the alsa-firmware package but
alsa-firmware-loaders is missing from somewhere. Hopefully not
needed.
2to3 is gone, hope it is not still needed.

Fixup Services

systemctl for a few things
Re-enable recipes
remove unneeded services
Cleanup other warnings and errors

Remove bad packages
Remove invalid packages
more network wifi packages..
Reorg and resolve packages
APT packages for resolution
Narrow down failing ffmpeg and mplayer
Add leading spaces to split lines to fix highligting
Remove dhcpcd-dbus missing apt package
Improve apt_list helper
Move old references from bookworm to trixie

Add -q to apt for better logging
Add Zynth_x86_64.md to hold instructions for zynthian x86_64
Streamline this a bit more to move x86_64 into it's own configs
Use MACHINE_HW_NAME to determine arch specific changes
Splitting out more x86_64, Use rpi version 86 to indicate
x86_64 stuff... This is consumed buy webconf
Stop removing apt packages
Something is removing clang before we compile all of the recipes
get rid of apt clean, and apt autoremove.
AMD64 symlink to i2c-0 so that scan does not fail.
Double check to see if it the i2c-0 exists before trying to link
Set env vars for keypad left for on-screen for x86_64
alias mkdir to add -p so that existing directories don't fail

Prebuilt will not work with x86_64, so use the setup_plugins
to compile more packages in recipes.
Modify setup to install drivers for Acer Laptop
Fix permissions on new x86_64 plugins

Fixup recipes for x86_64 add dsp56300 fork

Install missing clang compiler, and keep dsp56300 cloned dir
Setup x86_64 build for dsp56300/gearmulator
Pull in a fork, and add the presets for x86_64

Manually regen engines and presets
Possibly a failure is preventing the engines from being rebuilt.
Split these off to call the sbin/regenerate scripts individually.

Make it a single line setup command in the readme.

Streamline switching repos at launch to make development smoother

This sets the $ZYNTHIAN_BASE and $ZYNTHIAN_BRANCH when
setup_zynthian.sh is started, allowing for first time configuration
that is sticky.

Fixup ENVVAR names

Fix the GIT_BASE to ZYNTHIAN_BASE rename

Fix script running from /zynthian not ~

Move the initial script to /zynthian correct for bare "cd"

Add helper for git branch, and make branch optional

Use the repo default for branch unless it is specified.

Fix Fluidsynth, Dummy out LVTK for now, Remove Sooperlooper

This is failing somewhere basic in waf. Skip for now, and
hopefully it is not critical.

Something in the old waf, on debian 13 was crashing due to
invalid characters or something similar to the unpacking of the
executable. Update that here
Change sed version number and wildcard match the remainder in
hopes that this will patch up what is needed.

Fixup Fluidsynth service

Was throwing errors, possibly no longer auto installed/started

Remove Sooperlooper and mod-ui for now

sooperlooper has not made it yet to trixie, but is in forky
rather than pull in unstable, wait for it.
Avoiding mod-ui until I can get the base system working again

Dealing with more partial install recovery

SF2 ln command fails

setbfree default cfg

remove the old default so that the ln succeeds. Could also check
for its exist first.

Remove TAL-U-NO on amd64

The LV2 package is only available from zynthian, and it is only
for aarch64 (Arm Pi) Revisit this later if support can be added

Update Zynth_amd64.md

Updated instructions and hints for amd64 install

Fix recipe for liblo on Debian 13

Install from package, don't attempt to build
Only modify /etc/fstab for tmpfs once
Avoid messages about multiple entries

Remove anacron

Disable anacron
remove mod-ttymidi for amd64

Package early install openssh-server and remove apparmor

setup ssh so that I can ssh into the test VM a bit easier.
remove apparmor so that it does not get in the way.

Add SKIP_PLUGINS as envvar to setup script

invoke from command line
SKIP_PLUGINS=1 scripts/setup_zynthian.sh to skip plugin builds
which take a long time on the VM

Update Zynth_amd64.md docs slightly

Update to use wcurl command and more streamlined 2 line
command to kick off the full install

It ran into issues, partly with the clone names. This would need
more planning to actually work so it goes on the back burner.

Cleanup initial instructions and simplify proceedure

Use git for the initial install, its needed anyways and simplifies
the instructions.

+++ scripts/setup_zynthian.sh
@@ -7,10 +7,17 @@
 # instructions:
 #
 # 1. Run first time: sh ./setup_zynthian.sh
-# 2. Reboot: It should reboot automaticly after step 1
+# 2. Reboot: It should reboot automaticly after step 1 (rPi only).
 # 3. Run second time: screen -t setup -L sh ./setup_zynthian.sh
 # 4. Take a good beer, sit down and relax ... ;-)
 # 
+# Pass the Base Git repository URL as first parameter if you want to use
+# a different repository (for development purposes mainly) and the branch
+# name as an optional second parameter.
+# Example:
+#     ./setup_zynthian.sh https://github.com/your-fork/ your-branch
+#
+# if the second parameter is omitted, the repo default branch will be used.
 # Copyright (C) 2015-2024 Fernando Moyano <jofemodo@zynthian.org>
 #
 #******************************************************************************
@@ -28,22 +35,77 @@
 # For a full copy of the GNU General Public License see the LICENSE.txt file.
 # 
 #******************************************************************************
+mkdir -p /zynthian
+cd /zynthian
 
-cd
-
-if [ "$1" = "wiggle" ] || [ ! -f ~/.wiggled ]; then
-	echo `date` >  ~/.wiggled
-	raspi-config --expand-rootfs
-	reboot
-else
-	if [ ! -d "zynthian-sys" ]; then
-		apt-get update
-		apt-get -y install apt-utils git parted screen
-		git clone -b oram https://github.com/zynthian/zynthian-sys.git
+# write base_repo.sh if arguments were supplied
+if [ $# -gt 0 ]; then
+	echo "export ZYNTHIAN_BASE=${1:-https://github.com/zynthian}" > base_repo.sh
+	if [ $# -gt 1 ]; then
+		echo "export ZYNTHIAN_BRANCH=\"${2}\"" >> base_repo.sh
+	else
+		echo "export ZYNTHIAN_BRANCH=\"\"" >> base_repo.sh
 	fi
-	cd zynthian-sys/scripts
-	./setup_system_raspioslite_64bit_bookworm.sh
-	cd
-	rm -rf zynthian-sys
 fi
 
+if [ -f base_repo.sh ]; then
+	source base_repo.sh
+else
+	# default values
+	ZYNTHIAN_BASE="https://github.com/zynthian"
+	ZYNTHIAN_BRANCH=""
+fi
+
+# Helper function to return branch parameter
+has_branch() {
+	if [ -n "$1" ]; then
+		echo "-b $1"
+	fi
+}
+
+
+set -e #exit on error, so get it right!!!!!
+
+set -x #enable command tracing
+LOG_FILE="./setup_zynthian.log"
+
+echogreen() {
+	echo -e "\e[32m" $1 "\e[0m"
+}
+
+echogreen "Starting Zynthian Setup Script"
+
+# Redirect all stdout and stderr to the log file appending for when we reboot etc.
+exec > >(tee -a "${LOG_FILE}") 2>&1
+
+if uname -m | grep -qi x86_64; then
+	echo "x86_64 detected."
+	export IS_X86_64=true
+else
+	export IS_X86_64=false
+fi
+
+if [ "$1" = "wiggle" ] || ([ ! -f ~/.wiggled ] && [ "$IS_X86_64" = "false" ]); then
+	echo `date` > ~/.wiggled
+	raspi-config --expand-rootfs
+	reboot
+	exit 0
+fi
+
+apt-get update
+apt-get -q -y install apt-utils git parted screen
+
+if [ ! -d "zynthian-sys" ]; then
+	git clone $(has_branch ${ZYNTHIAN_BRANCH}) ${ZYNTHIAN_BASE}/zynthian-sys.git
+else
+	cd zynthian-sys
+	git pull
+	cd /zynthian
+fi
+
+cd zynthian-sys/scripts
+if [ "$IS_X86_64" = "true" ]; then
+	./setup_system_amd64_trixie.sh
+else
+	./setup_system_raspiolite_64bit_bookworm.sh
+fi
